requirements.txt
Flask==2.3.2
yfinance==0.2.21
pandas==2.2.2
